+++ src/main/java/http/common/HeaderFieldName.java
@@ -1,10 +1,11 @@
-package http;
+package http.common;
 
 public enum HeaderFieldName {
     LOCATION("Location"),
     CONTENT_TYPE("Content-Type"),
     CONTENT_LENGTH("Content-Length"),
     COOKIE("Cookie"),
+    SET_COOKIE("Set-Cookie"),
     ;
 
     private String value;
