+++ src/main/java/http/request/QueryString.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import org.apache.logging.log4j.util.Strings;
 import utils.UrlUtf8Decoder;
