+++ src/main/java/http/response/Header.java
@@ -1,6 +1,6 @@
 package http.response;
 
-import http.HeaderField;
+import http.common.HeaderField;
 import org.apache.logging.log4j.util.Strings;
 
 import java.util.HashMap;
