+++ src/test/java/http/CookiesTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.common.Cookies;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
