+++ src/test/java/http/ProtocolAndVersionTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.common.ProtocolAndVersion;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
